Rename imagenet_dataset.py => imagenet_dataset_tf.py
